requirements.txt
pandas
pinecone-client
sentence-transformers
torch
tqdm
pdfminer-six